refactor: move methods of GlobalExceptionHandler to the new GlobalExceptionHandler
