+++ src/main/java/com/postech/fastfood/config/GlobalExceptionHandler.java
@@ -3,17 +3,40 @@ package com.postech.fastfood.config;
 import com.postech.fastfood.core.exception.CpfAlreadyInUseException;
 import com.postech.fastfood.core.exception.CustomerNotFoundException;
 import com.postech.fastfood.core.exception.EmailAlreadyExistsException;
+import com.postech.fastfood.core.exception.FastFoodException;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import java.net.URI;
 import java.util.HashMap;
 import java.util.Map;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
+
+    private static final URI URI_VAZIA = URI.create("");
+
+    private ProblemDetail montarProblemDetail(String title, HttpStatus status, String message) {
+        final ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(status, message);
+        problemDetail.setType(URI_VAZIA);
+        problemDetail.setTitle(title);
+        return problemDetail;
+    }
+
+    @ExceptionHandler(FastFoodException.class)
+    public ProblemDetail fastFoodException(FastFoodException e) {
+        return montarProblemDetail(e.title, e.status, e.getLocalizedMessage());
+    }
+
+    @ExceptionHandler(AuthenticationException.class)
+    public ProblemDetail authenticationException(AuthenticationException e) {
+        return montarProblemDetail("Teste", HttpStatus.BAD_REQUEST, e.getLocalizedMessage());
+    }
 
     @ExceptionHandler(CustomerNotFoundException.class)
     public ResponseEntity<String> handleCustomerNotFound(CustomerNotFoundException ex) {
